refactor k8scluster
